requirements.txt
# UI e Framework
python-fasthtml
uvicorn
python-multipart
python-dotenv
streamlit
streamlit-ace
streamlit-antd-components
streamlit-cropper
defusedxml

# Gestione Immagini e Dati
Pillow>=10.0.0
pandas
requests>=2.31.0
beautifulsoup4
tqdm>=4.66.0

# PDF e OCR Engines
PyMuPDF
img2pdf>=0.5.1
openai
anthropic
